requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langgraph
faiss-cpu
